功能完善、优化

+++ page-list-wrapper/src/main/java/ws/dyt/pagelist/entity/IPageIndex.java
@@ -10,5 +10,5 @@ public interface IPageIndex {
      * 返回分页标识id
      * @return
      */
-    int getPageIndex();
+    long getPageIndex();
 }
